1) Moved Likelihood to other file 2) Added window mass constraint (as in ICL work) 3) Rewritten plotting module (moved making subplot to another function) for more flexibility in plotting 4) Added some points to fastMTT_test (tutorial) 5) Actualized readme
